@@ -12,44 +12,38 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.7",
-    "@angular/common": "^7.2.7",
-    "@angular/compiler": "^7.2.7",
-    "@angular/core": "^7.2.7",
-    "@angular/forms": "^7.2.7",
-    "@angular/http": "^7.2.7",
-    "@angular/platform-browser": "^7.2.7",
-    "@angular/platform-browser-dynamic": "^7.2.7",
-    "@angular/router": "^7.2.7",
-    "angular2-materialize": "^15.1.10",
-    "core-js": "^2.4.1",
+    "@angular/animations": "^17.3.0",
+    "@angular/common": "^17.3.12",
+    "@angular/compiler": "^17.3.0",
+    "@angular/core": "^17.3.12",
+    "@angular/forms": "^17.3.0",
+    "@angular/platform-browser": "^17.3.0",
+    "@angular/platform-browser-dynamic": "^17.3.0",
+    "@angular/router": "^17.3.0",
+    "@po-ui/ng-components": "^17.26.16",
+    "@po-ui/style": "^19.6.0",
+    "@totvs/protheus-lib-core": "^19.0.0",
+    "core-js": "^3.41.0",
     "hammerjs": "^2.0.8",
-    "jquery": "^3.2.1",
+    "jquery": "^3.7.1",
     "materialize-css": "^1.0.0",
-    "rxjs": "^6.4.0",
-    "rxjs-compat": "^6.0.0-rc.0",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.8.29"
+    "rxjs": "~7.8.0",
+    "subsink": "^1.0.2",
+    "tslib": "^2.8.1",
+    "web-animations-js": "^2.3.2",
+    "zone.js": "^0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.4",
-    "@angular/compiler-cli": "^7.2.7",
-    "@angular/language-service": "^7.2.7",
-    "@types/jasmine": "~2.8.6",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.0",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "~3.2.4"
+    "@angular-devkit/build-angular": "^17.3.14",
+    "@angular/cli": "^17.3.12",
+    "@angular/compiler-cli": "^17.3.0",
+    "@types/jasmine": "~5.1.0",
+    "jasmine-core": "~5.1.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "~5.4.2"
   }
 }
