@@ -1,7 +1,7 @@
 import { AlunosModule } from './../alunos.module';
 import { AlunosService } from './../alunos.service';
 import { ActivatedRoute, Router } from '@angular/router';
-import { Subscription } from 'rxjs/Rx';
+import { Subscription } from 'rxjs';
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Aluno } from './../aluno';
 
