@@ -1,18 +1,22 @@
-import { CursoNaoEncontradoComponent } from './cursos/curso-nao-encontrado/curso-nao-encontrado.component';
-import { CursoDetalheComponent } from './cursos/curso-detalhe/curso-detalhe.component';
 import { ModuleWithProviders } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
+
+import { LoginComponent } from './login/login.component';
 import { HomeComponent } from './home/home.component';
 import { CursosComponent } from './cursos/cursos.component';
-import { LoginComponent } from './login/login.component';
+import { CursoNaoEncontradoComponent } from './cursos/curso-nao-encontrado/curso-nao-encontrado.component';
+import { CursoDetalheComponent } from './cursos/curso-detalhe/curso-detalhe.component';
+import { AlunosComponent} from './alunos/alunos.component';
+
 
 const APP_ROUTES: Routes = [
-    { path: 'cursos', component: CursosComponent },
-    { path: 'curso/:id', component: CursoDetalheComponent },
-    { path: 'login', component: LoginComponent },
-    { path: 'naoEncontrado', component: CursoNaoEncontradoComponent },
-    { path: '', component: HomeComponent }
+  { path: '', component: LoginComponent },
+  { path: 'home', component: HomeComponent },
+  { path: 'cursos', component: CursosComponent },
+  { path: 'curso/:id', component: CursoDetalheComponent },
+  { path: 'alunos', component: AlunosComponent },
+  { path: 'naoEncontrado', component: CursoNaoEncontradoComponent },
 ];
 
-export const routing: ModuleWithProviders = RouterModule.forRoot(APP_ROUTES);
+export const routing: ModuleWithProviders<RouterModule> = RouterModule.forRoot(APP_ROUTES);
