@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { Subscription } from 'rxjs/Rx';
+import { Subscription } from 'rxjs';
 
 import { CursosService } from './cursos.service';
 
@@ -18,7 +18,7 @@ export class CursosComponent implements OnInit {
   constructor(
     private cursosService: CursosService,
     private route: ActivatedRoute,
-    private router: Router
+   private router: Router
   ) { }
 
   ngOnInit() {
@@ -35,10 +35,11 @@ export class CursosComponent implements OnInit {
     this.inscricao.unsubscribe();
   }
 
-  proximaPagina(){
-    //this.pagina++;
-    this.router.navigate(['/cursos'], 
-      {queryParams: {'pagina': ++this.pagina}});
-  }
+ proximaPagina(){
+    this.pagina++;
+    this.router.navigate(['/cursos'],
+     {queryParams: {'pagina': ++this.pagina}}
+    );
+ }
 
 }
