@@ -1,12 +1,12 @@
 import { IFormCanDeactivate } from './iform-candeactivate';
-import { Observable } from 'rxjs/Rx';
+import { Observable } from 'rxjs';
 import { ActivatedRouteSnapshot, RouterStateSnapshot, CanDeactivate } from '@angular/router';
 import { Injectable } from '@angular/core';
-import { AlunoFormComponent } from './../alunos/aluno-form/aluno-form.component';
+//import { AlunoFormComponent } from './../alunos/aluno-form/aluno-form.component';
 
 @Injectable()
 export class AlunosDeactivateGuard implements CanDeactivate<IFormCanDeactivate> {
-                
+
         canDeactivate(
             component: IFormCanDeactivate,
             route: ActivatedRouteSnapshot,
