@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs/Rx';
+import { Observable } from 'rxjs/';
 import { CanActivateChild, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 
 @Injectable()
@@ -10,13 +10,13 @@ export class AlunosGuard implements CanActivateChild {
             state: RouterStateSnapshot
         ): Observable<boolean>|boolean {
 
-            //console.log(route);
-            //console.log(state);
+            console.log(route);
+            console.log(state);
 
             console.log('AlunosGuard: Guarda de rota filha');
 
             if (state.url.includes('editar')){
-                //alert('Usuário sem acesso');
+                alert('Usuário sem acesso');
                 //return Observable.of(false);
             }
 
