@@ -1,12 +1,12 @@
-import { Component, OnInit } from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-meu-primeiro2',
+  standalone: true,
+  imports: [],
   templateUrl: './meu-primeiro2.component.html',
-  styleUrls: ['./meu-primeiro2.component.css']
+  styleUrl: './meu-primeiro2.component.css'
 })
 export class MeuPrimeiro2Component {
-
-
 
 }
