@@ -7,7 +7,7 @@ import { CursoDetalheComponent } from './curso-detalhe/curso-detalhe.component';
 import { CursoNaoEncontradoComponent } from './curso-nao-encontrado/curso-nao-encontrado.component';
 
 const cursosRoutes: Routes = [
-    { path: '', component: CursosComponent },
+    { path: 'cursos', component: CursosComponent },
     { path: 'naoEncontrado', component: CursoNaoEncontradoComponent },
     { path: ':id', component: CursoDetalheComponent }
 ];
