@@ -15,7 +15,7 @@ export class AuthService {
 
   fazerLogin(usuario: Usuario){
 
-    if (usuario.nome === 'usuario@email.com' && 
+    if (usuario.nome === 'usuario@email.com' &&
       usuario.senha === '123456') {
 
       this.usuarioAutenticado = true;
