@@ -13,7 +13,7 @@ export class CursosComponent implements OnInit {
   cursos: string[] = [];
   //cursosService: CursosService;
 
-  constructor(private cursosService: CursosService) { 
+  constructor(private cursosService: CursosService) {
     //this.cursosService = new CursosService();
     //this.cursosService = _cursosService;
   }
